--- particle_filter/modularized_particle_filter/include/modularized_particle_filter/correction/correction_util.hpp
@@ -1,36 +1,0 @@
-// Copyright 2023 TIER IV, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef MODULARIZED_PARTICLE_FILTER__CORRECTION__CORRECTION_UTIL_HPP_
-#define MODULARIZED_PARTICLE_FILTER__CORRECTION__CORRECTION_UTIL_HPP_
-
-#include "rclcpp/rclcpp.hpp"
-
-#include "modularized_particle_filter_msgs/msg/particle_array.hpp"
-
-#include <boost/circular_buffer.hpp>
-
-#include <optional>
-
-namespace yabloc
-{
-namespace modularized_particle_filter
-{
-std::optional<modularized_particle_filter_msgs::msg::ParticleArray> find_synced_particles(
-  boost::circular_buffer<modularized_particle_filter_msgs::msg::ParticleArray> circular_buffer,
-  rclcpp::Time time);
-}
-}  // namespace yabloc
-
-#endif  // MODULARIZED_PARTICLE_FILTER__CORRECTION__CORRECTION_UTIL_HPP_
