--- particle_filter/modularized_particle_filter/launch/run.launch.xml
@@ -1,20 +1,0 @@
-<launch>
-    <!-- ========================================= -->
-    <arg name="use_sim_time" default="true"/>
-    <arg name="log_level" default="warn"/>
-
-    <!-- ========================================= -->
-    <node pkg="modularized_particle_filter" exec="predictor" name="predictor" output="screen" args="--ros-args --log-level warn">
-        <remap from="twist" to="/kalman/twist" />
-        <remap from="initialpose" to="/initialpose" />
-        <remap from="pose" to="/init/pose"/> 
-
-        <param name="use_sim_time" value="$(var use_sim_time)"/>
-        <param name="num_of_particles" value="500" />
-        <param name="resampling_interval_seconds" value="1.5" />
-        <param name="prediction_rate" value="50.0" />
-
-        <param name="static_linear_covariance" value="0.09" />
-        <param name="static_angular_covariance" value="0.01" />
-    </node>
-</launch>
