--- particle_filter/modularized_particle_filter/package.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>modularized_particle_filter</name>
-  <version>0.1.0</version>
-  <description>The modularized_particle_filter package</description>
-  <maintainer email="kento.yabuuchi.2@tier4.jp">Kento Yabuuchi</maintainer>
-  <license>Apache License 2.0</license>
-
-  <buildtool_depend>ament_cmake_ros</buildtool_depend>
-  <buildtool_depend>rosidl_default_generators</buildtool_depend>
-
-  <depend>rclcpp</depend>
-  <depend>tf2</depend>
-  <depend>tf2_ros</depend>
-  <depend>std_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>sensor_msgs</depend>
-  <depend>visualization_msgs</depend>
-  <depend>tf2_sensor_msgs</depend>
-  <depend>tf2_geometry_msgs</depend>
-
-  <depend>sophus</depend>
-  <depend>yabloc_common</depend>
-  <depend>modularized_particle_filter_msgs</depend>
-
-  <test_depend>ament_cmake_gtest</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
