--- particle_filter/modularized_particle_filter/src/prediction/predictor_node.cpp
@@ -1,28 +1,0 @@
-// Copyright 2023 TIER IV, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "modularized_particle_filter/prediction/predictor.hpp"
-
-#include <rclcpp/rclcpp.hpp>
-
-#include <memory>
-
-int main(int argc, char ** argv)
-{
-  namespace mpf = yabloc::modularized_particle_filter;
-  rclcpp::init(argc, argv);
-  rclcpp::spin(std::make_shared<mpf::Predictor>());
-  rclcpp::shutdown();
-  return 0;
-}
