--- particle_filter/modularized_particle_filter_msgs/package.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>modularized_particle_filter_msgs</name>
-  <version>0.1.0</version>
-  <description>The modularized_particle_filter_msgs package</description>
-  <maintainer email="kento.yabuuchi.2@tier4.jp">Kento Yabuuchi</maintainer>
-  <license>Apache License 2.0</license>
-
-  <buildtool_depend>ament_cmake_ros</buildtool_depend>
-  <buildtool_depend>rosidl_default_generators</buildtool_depend>
-
-  <depend>std_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>sensor_msgs</depend>
-
-  <exec_depend>rosidl_default_runtime</exec_depend>
-  <member_of_group>rosidl_interface_packages</member_of_group>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
